// src/state/types/actionEnums.ts
export enum ActionEnums {
  // Loading
  AUTH_IS_LOADING = "AUTH_IS_LOADING",
  ADMIN_IS_LOADING = "ADMIN_IS_LOADING",
  SERVICES_IS_LOADING = "SERVICES_IS_LOADING",
  CHAT_IS_LOADING = "CHAT_IS_LOADING",

  /// Clean state
  CLEAN_AUTH_STATE = "CLEAN_AUTH_STATE",

  //////
  SIGN_IN_SUCCESS = "SIGN_IN_SUCCESS",
  SIGN_IN_FAIL = "SIGN_IN_FAIL",
  SIGNOUT_SUCCESS = "SIGNOUT_SUCCESS",
  SIGNOUT_FAIL = "SIGNOUT_FAIL",
  SIGN_UP_SUCCESS = "SIGN_UP_SUCCESS",
  SIGN_UP_FAIL = "SIGN_UP_FAIL",
  GET_USER_SUCCESS = "GET_USER_SUCCESS",
  GET_USER_FAIL = "GET_USER_FAIL",
  GET_USER_DETAILS_SUCCESS = "GET_USER_DETAILS_SUCCESS",
  GET_USER_DETAILS_FAIL = "GET_USER_DETAILS_FAIL",
  ADMIN_GET_ALL_USERS_SUCCESS = "ADMIN_GET_ALL_USERS_SUCCESS",
  ADMIN_GET_ALL_USERS_FAIL = "ADMIN_GET_ALL_USERS_FAIL",
  ADMIN_UPDATE_USER_SUCCESS = "ADMIN_UPDATE_USER_SUCCESS",
  ADMIN_UPDATE_USER_FAIL = "ADMIN_UPDATE_USER_FAIL",
  ADMIN_DELETE_USER_SUCCESS = "ADMIN_DELETE_USER_SUCCESS",
  ADMIN_DELETE_USER_FAIL = "ADMIN_DELETE_USER_FAIL",
  SEND_OTP_SUCCESS = "SEND_OTP_SUCCESS",
  SEND_OTP_FAIL = "SEND_OTP_FAIL",
  VERIFY_OTP_SUCCESS = "VERIFY_OTP_SUCCESS",
  VERIFY_OTP_FAIL = "VERIFY_OTP_FAIL",
  GET_SOCIAL_STATE_SUCCESS = "GET_SOCIAL_STATE_SUCCESS",
  GET_SOCIAL_STATE_FAIL = "GET_SOCIAL_STATE_FAIL",
  SOCIAL_LOGIN_SUCCESS = "SOCIAL_LOGIN_SUCCESS",
  SOCIAL_LOGIN_FAIL = "SOCIAL_LOGIN_FAIL",
  CONFIRM_EMAIL_SUCCESS = "CONFIRM_EMAIL_SUCCESS",
  CONFIRM_EMAIL_FAIL = "CONFIRM_EMAIL_FAIL",
  REQUEST_RESET_PASSWORD_SUCCESS = "REQUEST_RESET_PASSWORD_SUCCESS",
  REQUEST_RESET_PASSWORD_FAIL = "REQUEST_RESET_PASSWORD_FAIL",
  RESET_PASSWORD_SUCCESS = "RESET_PASSWORD_SUCCESS",
  RESET_PASSWORD_FAIL = "RESET_PASSWORD_FAIL",
  RESET_PASSWORD_OTP_SUCCESS = "RESET_PASSWORD_OTP_SUCCESS",
  RESET_PASSWORD_OTP_FAIL = "RESET_PASSWORD_OTP_FAIL",
  UPDATE_USER_SUCCESS = "UPDATE_USER_SUCCESS",
  UPDATE_USER_FAIL = "UPDATE_USER_FAIL",
  GET_CATEGORIES_SUCCESS = "GET_CATEGORIES_SUCCESS",
  GET_CATEGORIES_FAIL = "GET_CATEGORIES_FAIL",
  GET_SUB_CATEGORIES_SUCCESS = "GET_SUB_CATEGORIES_SUCCESS",
  GET_SUB_CATEGORIES_FAIL = "GET_SUB_CATEGORIES_FAIL",
  GET_WILAYAS_SUCCESS = "GET_WILAYAS_SUCCESS",
  GET_WILAYAS_FAIL = "GET_WILAYAS_FAIL",
  GET_COMMUNES_SUCCESS = "GET_COMMUNES_SUCCESS",
  GET_COMMUNES_FAIL = "GET_COMMUNES_FAIL",
  DELETE_PREVIOUS_WORK_PHOTO_SUCCESS = "DELETE_PREVIOUS_WORK_PHOTO_SUCCESS",
  DELETE_PREVIOUS_WORK_PHOTO_FAIL = "DELETE_PREVIOUS_WORK_PHOTO_FAIL",
  ADMIN_UPDATE_CATEGORY_SUCCESS = "ADMIN_UPDATE_CATEGORY_SUCCESS",
  ADMIN_UPDATE_CATEGORY_FAIL = "ADMIN_UPDATE_CATEGORY_FAIL",
  ADMIN_CREATE_CATEGORY_SUCCESS = "ADMIN_CREATE_CATEGORY_SUCCESS",
  ADMIN_CREATE_CATEGORY_FAIL = "ADMIN_CREATE_CATEGORY_FAIL",
  ADMIN_UPDATE_SUB_CATEGORY_SUCCESS = "ADMIN_UPDATE_SUB_CATEGORY_SUCCESS",
  ADMIN_UPDATE_SUB_CATEGORY_FAIL = "ADMIN_UPDATE_SUB_CATEGORY_FAIL",
  ADMIN_CREATE_SUB_CATEGORY_SUCCESS = "ADMIN_CREATE_SUB_CATEGORY_SUCCESS",
  ADMIN_CREATE_SUB_CATEGORY_FAIL = "ADMIN_CREATE_SUB_CATEGORY_FAIL",
  ADMIN_DELETE_SUB_CATEGORY_SUCCESS = "ADMIN_DELETE_SUB_CATEGORY_SUCCESS",
  ADMIN_DELETE_SUB_CATEGORY_FAIL = "ADMIN_DELETE_SUB_CATEGORY_FAIL",
  ADMIN_DELETE_CATEGORY_SUCCESS = "ADMIN_DELETE_CATEGORY_SUCCESS",
  ADMIN_DELETE_CATEGORY_FAIL = "ADMIN_DELETE_CATEGORY_FAIL",
  ADMIN_CREATE_ADMIN_SUCCESS = "ADMIN_CREATE_ADMIN_SUCCESS",
  ADMIN_CREATE_ADMIN_FAIL = "ADMIN_CREATE_ADMIN_FAIL",
  CLIENT_GET_JOBS_SUCCESS = "CLIENT_GET_JOBS_SUCCESS",
  CLIENT_GET_JOBS_FAIL = "CLIENT_GET_JOBS_FAIL",
  ARTISAN_GET_JOBS_SUCCESS = "ARTISAN_GET_JOBS_SUCCESS",
  ARTISAN_GET_JOBS_FAIL = "ARTISAN_GET_JOBS_FAIL",
  ARTISAN_GET_PREVIOUS_JOBS_SUCCESS = "ARTISAN_GET_PREVIOUS_JOBS_SUCCESS",
  ARTISAN_GET_PREVIOUS_JOBS_FAIL = "ARTISAN_GET_PREVIOUS_JOBS_FAIL",
  GET_CHAT_ROOMS_SUCCESS = "GET_CHAT_ROOMS_SUCCESS",
  GET_CHAT_ROOMS_FAIL = "GET_CHAT_ROOMS_FAIL",
  GET_NOTIFICATIONS_SUCCESS = "GET_NOTIFICATIONS_SUCCESS",
  GET_NOTIFICATIONS_FAIL = "GET_NOTIFICATIONS_FAIL",
  MARK_NOTIFICATIONS_AS_READ_SUCCESS = "MARK_NOTIFICATIONS_AS_READ_SUCCESS",
  MARK_NOTIFICATIONS_AS_READ_FAIL = "MARK_NOTIFICATIONS_AS_READ_FAIL",
  SWITCH_TO_CLIENT = "SWITCH_TO_CLIENT",
  CREATE_JOB_SUCCESS = "CREATE_JOB_SUCCESS",
  CREATE_JOB_FAIL = "CREATE_JOB_FAIL",
  DELETE_JOB_SUCCESS = "DELETE_JOB_SUCCESS",
  DELETE_JOB_FAIL = "DELETE_JOB_FAIL",
  ADD_JOB_IMAGES_SUCCESS = "ADD_JOB_IMAGES_SUCCESS",
  ADD_JOB_IMAGES_FAIL = "ADD_JOB_IMAGES_FAIL",
  ADMIN_GET_PAYMENT_RECEIPT_SUCCESS = "ADMIN_GET_PAYMENT_RECEIPT_SUCCESS",
  ADMIN_GET_PAYMENT_RECEIPT_FAIL = "ADMIN_GET_PAYMENT_RECEIPT_FAIL",
  ADMIN_UPDATE_PAYMENT_RECEIPT_SUCCESS = "ADMIN_UPDATE_PAYMENT_RECEIPT_SUCCESS",
  ADMIN_UPDATE_PAYMENT_RECEIPT_FAIL = "ADMIN_UPDATE_PAYMENT_RECEIPT_FAIL",
  ADMIN_GET_JOBS_SUCCESS = "ADMIN_GET_JOBS_SUCCESS",
  ADMIN_GET_JOBS_FAIL = "ADMIN_GET_JOBS_FAIL",
  SET_USER_DATA_FETCHED_FROM_TOKEN_TRUE = "SET_USER_DATA_FETCHED_FROM_TOKEN_TRUE",
}